+++ src/baseconfig.h
@@ -6,11 +6,23 @@
 #include <StreamUtils.h>
 #include <iomanip>  // needed by setw / setfill
 #include <sstream>
+#include <vectorlist.h>
 #include <_Release.h>
 
 class BaseConfig {
 
   public:
+
+    // Speaking identifiers
+    enum class GpioIdentifier : uint8_t {
+      BASECONFIG = 0,
+      FLOWCONTROL,
+      SENSOR,
+      VALVES,
+      ETH,
+      OTHER
+    };
+
     BaseConfig(fs::LittleFSFS& configFS);
     void      LoadJsonConfig();
 
@@ -58,7 +70,11 @@ class BaseConfig {
     const uint8_t&  GetMaxThreads()   const {return max_threads;}
 
     size_t          getFragmentation();
-     
+
+  // Maintains a list of currently 'reserved' GPIOs (SDA, SCL, 1Wire, Serial, etc.)
+
+  vectorlist<uint8_t, GpioIdentifier> disabledGPIO;
+
   private:
     fs::LittleFSFS configFS; 
     String    mqtt_server;
@@ -87,6 +103,7 @@ class BaseConfig {
     uint8_t   max_threads;
 
     std::function<void(const char*)> onLogValuesCallback; // Callback function pointer
+
 };
 
 extern BaseConfig* Config;